backend/requirement.txt
fastapi
uvicorn
pymysql
passlib[bcrypt]
python-jose
python-dotenv
httpx